requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
EasyProcess==1.1
entrypoint2==1.1
idna==3.4
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
pyunpack==0.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.2
scikit-learn==1.2.2
mlflow==2.3.2
-e .